README.md
Name:J.Brinil John
Reg.No:22122024
Email:brinil.john@msds.christuniversity.in
Class:2 MscDs A

LAB-1:Question:
Write a Java Program that will collect an employee's basic details that fall into different data types and displays them.
After the details have been displayed, with the help of conditional statements, check if the gender of the employee is 'm' or 'f'. 
It should print "MALE" for 'm' and "FEMALE" for 'f'.
Assume that you can divide the states among India into different regions (North, South, Central, East, and West). If the employee
is from the southern part of India, with the help of a switch statement, it shoulddisplay "The Employee is from the southern 
states of India; Preferable work location is in <state>", along with the basic details.Identify if the employee belongs to the 
top MNC Companies (Facebook, Google, Microsoft, Samsung, IBM, Apple); if so, print a message"The employee is working in Top 
MNC Companies".